backend/requirements.txt
flask
flask-cors
gunicorn
pymongo
python-dotenv
google-genai
